requirements.txt
Django==2.1.5
django-shortuuidfield==0.1.3
djangorestframework==3.9.2
djangorestframework-stubs==0.1.0
Pillow==5.4.1
pytz==2018.9
shortuuid==0.5.0
six==1.12.0
